+++ pkg/generator/generator.go
@@ -2,13 +2,18 @@ package generator
 
 import (
 	"bytes"
+	"context"
 	"html/template"
 	"path"
+	"regexp"
+	"strings"
 
 	"github.com/Masterminds/sprig"
 	"github.com/ghodss/yaml"
 	"github.com/giantswarm/microerror"
 	pathmodifier "github.com/giantswarm/valuemodifier/path"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 /*
@@ -37,12 +42,19 @@ import (
 						configmap-values.yaml.patch.template
 */
 
+var (
+	multipleDashPattern     = regexp.MustCompile("-{2,}")
+	invalidCharacterPattern = regexp.MustCompile("[^a-z0-9]+")
+)
+
 type Config struct {
-	Fs Filesystem
+	Fs               Filesystem
+	DecryptTraverser DecryptTraverser
 }
 
 type Generator struct {
-	fs Filesystem
+	fs               Filesystem
+	decryptTraverser DecryptTraverser
 }
 
 func New(config *Config) (*Generator, error) {
@@ -50,7 +62,8 @@ func New(config *Config) (*Generator, error) {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Fs must not be empty", config)
 	}
 	g := Generator{
-		fs: config.Fs,
+		fs:               config.Fs,
+		decryptTraverser: config.DecryptTraverser,
 	}
 
 	return &g, nil
@@ -67,11 +80,13 @@ func New(config *Config) (*Generator, error) {
 // 4. Patch global template (result of 2.) with installation-specific (result
 //    of 3.) app overrides (if available)
 // 5. Get global secrets template data
-// 6. Get installation-specific secrets template for the app (if available) and
+// 6. Decrypt global secrets if DecryptTraverser has been provided.
+// 7. Get installation-specific secrets template for the app (if available) and
 //    render it with installation secrets template data (result of 5.)
-func (g Generator) GenerateConfig(installation, app string) (configmap string, secrets string, err error) {
+func (g Generator) GenerateRawConfig(ctx context.Context, installation, app string) (configmap string, secrets string, err error) {
 	// 1.
 	configmapContext, err := g.getWithPatchIfExists(
+		ctx,
 		"default/config.yaml",
 		"installations/"+installation+"/config.yaml.patch",
 	)
@@ -81,6 +96,7 @@ func (g Generator) GenerateConfig(installation, app string) (configmap string, s
 
 	// 2.
 	configmapBase, err := g.getRenderedTemplate(
+		ctx,
 		"default/apps/"+app+"/configmap-values.yaml.template",
 		configmapContext,
 	)
@@ -92,7 +108,7 @@ func (g Generator) GenerateConfig(installation, app string) (configmap string, s
 	var configmapPatch string
 	{
 		filepath := "installations/" + installation + "/apps/" + app + "/configmap-values.yaml.patch.template"
-		patch, err := g.getRenderedTemplate(filepath, configmapContext)
+		patch, err := g.getRenderedTemplate(ctx, filepath, configmapContext)
 		if IsNotFound(err) {
 			configmapPatch = ""
 		} else if err != nil {
@@ -104,6 +120,7 @@ func (g Generator) GenerateConfig(installation, app string) (configmap string, s
 
 	// 4.
 	configmap, err = applyPatch(
+		ctx,
 		[]byte(configmapBase),
 		[]byte(configmapPatch),
 	)
@@ -113,6 +130,7 @@ func (g Generator) GenerateConfig(installation, app string) (configmap string, s
 
 	// 5.
 	secretsContext, err := g.getWithPatchIfExists(
+		ctx,
 		"installations/"+installation+"/secrets.yaml",
 		"",
 	)
@@ -121,7 +139,17 @@ func (g Generator) GenerateConfig(installation, app string) (configmap string, s
 	}
 
 	// 6.
+	if g.decryptTraverser != nil {
+		decryptedBytes, err := g.decryptTraverser.Traverse(ctx, []byte(secretsContext))
+		if err != nil {
+			return "", "", microerror.Mask(err)
+		}
+		secretsContext = string(decryptedBytes)
+	}
+
+	// 7.
 	secretsTemplate, err := g.getWithPatchIfExists(
+		ctx,
 		"default/apps/"+app+"/secret-values.yaml.template",
 		"",
 	)
@@ -131,9 +159,46 @@ func (g Generator) GenerateConfig(installation, app string) (configmap string, s
 		return "", "", microerror.Mask(err)
 	}
 
-	secrets, err = g.renderTemplate(secretsTemplate, secretsContext)
+	secrets, err = g.renderTemplate(ctx, secretsTemplate, secretsContext)
 	if err != nil {
 		return "", "", microerror.Mask(err)
+	}
+
+	return configmap, secrets, nil
+}
+
+func (g Generator) GenerateConfig(ctx context.Context, installation, app, ref string) (configmap *corev1.ConfigMap, secrets *corev1.Secret, err error) {
+	cm, s, err := g.GenerateRawConfig(ctx, installation, app)
+	if err != nil {
+		return nil, nil, microerror.Mask(err)
+	}
+
+	name := generateResourceName(app, ref)
+	meta := metav1.ObjectMeta{
+		Name:      name,
+		Namespace: "giantswarm",
+	}
+
+	configmap = &corev1.ConfigMap{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "ConfigMap",
+			APIVersion: "v1",
+		},
+		ObjectMeta: meta,
+		Data: map[string]string{
+			"configmap-values.yaml": cm,
+		},
+	}
+
+	secrets = &corev1.Secret{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Secret",
+			APIVersion: "v1",
+		},
+		ObjectMeta: meta,
+		Data: map[string][]byte{
+			"secret-values.yaml": []byte(s),
+		},
 	}
 
 	return configmap, secrets, nil
@@ -142,7 +207,7 @@ func (g Generator) GenerateConfig(installation, app string) (configmap string, s
 // getWithPatchIfExists provides contents of filepath overwritten by patch at
 // patchFilepath. File at patchFilepath may be non-existent, resulting in pure
 // file at filepath being returned.
-func (g Generator) getWithPatchIfExists(filepath, patchFilepath string) (string, error) {
+func (g Generator) getWithPatchIfExists(ctx context.Context, filepath, patchFilepath string) (string, error) {
 	var err error
 
 	var base []byte
@@ -169,20 +234,20 @@ func (g Generator) getWithPatchIfExists(filepath, patchFilepath string) (string,
 		}
 	}
 
-	result, err := applyPatch(base, patch)
+	result, err := applyPatch(ctx, base, patch)
 	if err != nil {
 		return "", microerror.Mask(err)
 	}
 	return result, nil
 }
 
-func (g Generator) getRenderedTemplate(filepath, templateData string) (string, error) {
+func (g Generator) getRenderedTemplate(ctx context.Context, filepath, templateData string) (string, error) {
 	templateBytes, err := g.fs.ReadFile(filepath)
 	if err != nil {
 		return "", microerror.Mask(err)
 	}
 
-	result, err := g.renderTemplate(string(templateBytes), templateData)
+	result, err := g.renderTemplate(ctx, string(templateBytes), templateData)
 	if err != nil {
 		return "", microerror.Mask(err)
 	}
@@ -190,7 +255,7 @@ func (g Generator) getRenderedTemplate(filepath, templateData string) (string, e
 	return result, nil
 }
 
-func applyPatch(base, patch []byte) (string, error) {
+func applyPatch(ctx context.Context, base, patch []byte) (string, error) {
 	var basePathSvc *pathmodifier.Service
 	{
 		c := pathmodifier.DefaultConfig()
@@ -238,7 +303,7 @@ func applyPatch(base, patch []byte) (string, error) {
 	return string(outputBytes), nil
 }
 
-func (g Generator) renderTemplate(templateText string, templateData string) (string, error) {
+func (g Generator) renderTemplate(ctx context.Context, templateText string, templateData string) (string, error) {
 	c := map[string]interface{}{}
 	err := yaml.Unmarshal([]byte(templateData), &c)
 	if err != nil {
@@ -264,7 +329,7 @@ func (g Generator) renderTemplate(templateText string, templateData string) (str
 }
 
 func (g Generator) include(templateName string, templateData interface{}) (string, error) {
-	contents, err := g.fs.ReadFile(path.Join("include", templateName+".yaml"))
+	contents, err := g.fs.ReadFile(path.Join("include", templateName+".yaml.template"))
 	if err != nil {
 		return "", microerror.Mask(err)
 	}
@@ -281,4 +346,15 @@ func (g Generator) include(templateName string, templateData interface{}) (strin
 	}
 
 	return out.String(), nil
+}
+
+func generateResourceName(elements ...string) string {
+	name := strings.Join(elements, "-")
+	name = string(invalidCharacterPattern.ReplaceAll([]byte(name), []byte("-")))
+	name = string(multipleDashPattern.ReplaceAll([]byte(name), []byte("-")))
+	name = strings.ToLower(strings.Trim(name, "-"))
+	if len(name) > 63 {
+		name = name[:63]
+	}
+	return name
 }