+++ pkg/generator/spec.go
@@ -1,8 +1,15 @@
 package generator
 
-import "os"
+import (
+	"context"
+	"os"
+)
 
 type Filesystem interface {
 	ReadFile(string) ([]byte, error)
 	ReadDir(string) ([]os.FileInfo, error)
+}
+
+type DecryptTraverser interface {
+	Traverse(context.Context, []byte) ([]byte, error)
 }